+++ vaannotate/vaannotate_ai_backend/engine.py
@@ -64,10 +64,10 @@ from .services import (
 from .pipelines.active_learning import ActiveLearningPipeline
 from .label_configs import EMPTY_BUNDLE, LabelConfigBundle
 from .llm_backends import build_llm_backend
-from .core.retrieval import RetrievalCoordinator, SemanticQuery
-from .utils.hashing import _stable_rules_hash, stable_hash_pair, stable_hash_str
+from .services.rag_retriever import RAGRetriever
+from .services.disagreement_expander import DisagreementExpander
 from .utils.io import atomic_write_bytes, normalize01, read_table, write_table
-from .utils.jsonish import _jsonify_cols, _maybe_parse_jsonish
+from .utils.jsonish import _jsonify_cols
 from .utils.runtime import (
     CancelledError,
     LOGGER,
@@ -75,939 +75,11 @@ from .utils.runtime import (
     check_cancelled,
     iter_with_bar,
 )
-
-try:
-    from cachetools import LRUCache
-except Exception:
-    from collections import OrderedDict
-    class LRUCache(dict):
-        def __init__(self, maxsize=10000):
-            super().__init__(); self._order = OrderedDict(); self._max = maxsize
-        def __setitem__(self, k, v):
-            if k in self._order: self._order.move_to_end(k)
-            self._order[k] = None
-            super().__setitem__(k, v)
-            if len(self._order) > self._max:
-                old, _ = self._order.popitem(last=False)
-                super().pop(old, None)
-        def get(self, k, default=None):
-            if k in self._order: self._order.move_to_end(k)
-            return super().get(k, default)
         
 """
 Note: Embedding/index primitives (IndexConfig, Models, EmbeddingStore) now live
 in ``core.embeddings``.
 """
-
-class DisagreementExpander:
-    def __init__(
-        self,
-        cfg: DisagreementConfig,
-        repo: DataRepository,
-        retriever: RAGRetriever,
-        label_config_bundle: LabelConfigBundle | None = None,
-        llmfirst_cfg: LLMFirstConfig | None = None,
-        context_builder: ContextBuilder | None = None,
-    ):
-        self.cfg = cfg
-        self.repo = repo
-        self.retriever = retriever
-        self.label_config_bundle = label_config_bundle or EMPTY_BUNDLE
-        self._dependency_cache: Dict[str, tuple[dict, dict, list]] = {}
-        self.llmfirst_cfg = llmfirst_cfg
-        self.context_builder = context_builder
-
-    def _dependencies_for(self, labelset_id: Optional[str]) -> tuple[dict, dict, list]:
-        key = str(labelset_id or "__current__")
-        if key not in self._dependency_cache:
-            config = self.label_config_bundle.config_for_labelset(labelset_id)
-            try:
-                deps = build_label_dependencies(config)
-            except Exception:
-                deps = ({}, {}, [])
-            self._dependency_cache[key] = deps
-        return self._dependency_cache[key]
-
-    def high_entropy_seeds(self) -> pd.DataFrame:
-        dis = self.repo.reviewer_disagreement(round_policy=self.cfg.round_policy, decay_half_life=self.cfg.decay_half_life)
-        seeds = dis[dis["disagreement_score"] >= self.cfg.high_entropy_threshold].copy()
-        types = self.repo.label_types()
-        hard_df = self.repo.hard_disagree(types, date_days=int(self.cfg.date_disagree_days), num_abs=float(self.cfg.numeric_disagree_abs), num_rel=float(self.cfg.numeric_disagree_rel))
-        if not hard_df.empty:
-            hard_pairs = set(zip(hard_df.loc[hard_df['hard_disagree'],'unit_id'].astype(str), hard_df.loc[hard_df['hard_disagree'],'label_id'].astype(str)))
-            if hard_pairs:
-                add = dis[[ (str(r.unit_id), str(r.label_id)) in hard_pairs for r in dis.itertuples(index=False) ]].copy()
-                seeds = pd.concat([seeds, add], ignore_index=True).drop_duplicates(subset=['unit_id','label_id'])
-        seeds = seeds.sort_values("disagreement_score", ascending=False)
-
-        # ---- Parent→child gating on seeds (use prior-round consensus) ----
-        seeds = seeds.copy()
-        if "labelset_id" not in seeds.columns:
-            seeds["labelset_id"] = seeds.apply(
-                lambda r: self.repo.labelset_for_annotation(
-                    r.get("unit_id", ""),
-                    r.get("label_id", ""),
-                    round_id=r.get("round_id"),
-                )
-                or "",
-                axis=1,
-            )
-        else:
-            seeds["labelset_id"] = seeds["labelset_id"].fillna("").astype(str)
-        seeds["round_id"] = seeds.get("round_id", "").astype(str)
-        types = self.repo.label_types()
-        consensus = self.repo.last_round_consensus()  # {(unit_id,label_id)-> value str}
-
-        def _gate_seed(row: pd.Series) -> bool:
-            uid = str(row["unit_id"])
-            lid = str(row["label_id"])
-            labelset_id = str(row.get("labelset_id") or "").strip() or None
-            parent_to_children, child_to_parents, roots = self._dependencies_for(labelset_id)
-            roots_set = set(str(x) for x in (roots or []))
-            # Parents (roots) are always eligible; children need parent gate pass
-            if str(lid) in roots_set:
-                return True
-            parents = child_to_parents.get(str(lid), [])
-            if not parents:
-                return True  # not marked as child → treat as eligible
-            parent_preds = {(str(uid), str(p)): consensus.get((str(uid), str(p)), None) for p in parents}
-            # IMPORTANT: evaluate by prior-round consensus; robust evaluator handles casing/types
-            config = self.label_config_bundle.config_for_labelset(labelset_id)
-            return evaluate_gating(str(lid), str(uid), parent_preds, types, config)
-
-        if not seeds.empty:
-            seeds["unit_id"] = seeds["unit_id"].astype(str)
-            seeds["label_id"] = seeds["label_id"].astype(str)
-            seeds = seeds[seeds.apply(_gate_seed, axis=1)]
-        
-        rows = []
-        for lid, grp in seeds.groupby("label_id"):
-            rows.append(grp.head(self.cfg.seeds_per_label))
-        return pd.concat(rows, ignore_index=True) if rows else seeds.head(0)
-
-    def seed_snippets(self, unit_id: str, label_id: str, label_rules: str) -> List[str]:
-        spans = self.repo.get_prior_rationales(unit_id, label_id)
-        snips = [sp.get("snippet") for sp in spans if isinstance(sp,dict) and sp.get("snippet")]
-        if snips: return snips[: self.cfg.snippets_per_seed]
-        if self.context_builder is not None:
-            ctx = self.context_builder.build_context_for_label(
-                unit_id,
-                label_id,
-                label_rules,
-                topk_override=self.cfg.snippets_per_seed,
-                single_doc_context_mode=getattr(self.llmfirst_cfg, "single_doc_context", "rag"),
-                full_doc_char_limit=getattr(self.llmfirst_cfg, "single_doc_full_context_max_chars", None),
-            )
-        else:
-            ctx = self.retriever.retrieve_for_patient_label(
-                unit_id,
-                label_id,
-                label_rules,
-                topk_override=self.cfg.snippets_per_seed,
-                min_k_override=None,
-                mmr_lambda_override=None,
-            )
-        return [c["text"] for c in ctx if isinstance(c.get("text"), str)]
-
-    def expand(self, rules_map: Dict[str,str], seen_pairs: set) -> pd.DataFrame:
-        seeds = self.high_entropy_seeds()
-        print('disagreement seeds: ', seeds)
-        rows = []
-        for lid, grp in seeds.groupby("label_id"):
-            labelset_value = ""
-            if "labelset_id" in grp.columns:
-                non_empty = grp["labelset_id"].dropna().astype(str).str.strip()
-                if not non_empty.empty:
-                    labelset_value = str(non_empty.iloc[0])
-            round_value = ""
-            if "round_id" in grp.columns:
-                round_non_empty = grp["round_id"].dropna().astype(str).str.strip()
-                if not round_non_empty.empty:
-                    round_value = str(round_non_empty.iloc[0])
-            snips = []
-            for r in grp.itertuples(index=False):
-                snips.extend(self.seed_snippets(r.unit_id, lid, rules_map.get(lid,"")))
-            snips = snips[: self.cfg.seeds_per_label * self.cfg.snippets_per_seed]
-            if not snips: continue
-            cand = self.retriever.expand_from_snippets(lid, snips, seen_pairs, per_seed_k=self.cfg.similar_chunks_per_seed)
-            items = sorted(cand.items(), key=lambda kv: kv[1], reverse=True)[: self.cfg.expanded_per_label]
-            for uid, sc in items:
-                rows.append({
-                    "unit_id": uid,
-                    "label_id": lid,
-                    "score": float(sc),
-                    "bucket": "disagreement_expanded",
-                    "labelset_id": labelset_value,
-                    "round_id": round_value,
-                })
-        df = pd.DataFrame(rows)
-        if df.empty: return df
-        out = []
-        for lid, g in df.groupby("label_id"):
-            arr = g["score"].to_numpy()
-            g = g.copy(); g["score_n"] = normalize01(arr); out.append(g)
-        return pd.concat(out, ignore_index=True)
-
-
-class RAGRetriever:
-    _RR_CACHE_MAX = 200000
-    def __init__(self, store: EmbeddingStore, models: Models, cfg: RAGConfig, label_configs: Optional[dict]=None, notes_by_doc: Optional[Dict[str,str]]=None, repo: Optional[DataRepository]=None):
-        self.store = store; self.models = models; self.cfg = cfg
-        self.label_configs = label_configs or {}
-        self._notes_by_doc = notes_by_doc or {}
-        self._repo = repo
-        self._rr_cache = LRUCache(maxsize=self._RR_CACHE_MAX)
-        self._bm25_cache: Dict[str, dict] = {}
-        self._label_query_texts = {}   # (label_id, rules_hash, K) -> List[str]
-        self._label_query_embs  = {}   # (label_id, rules_hash, K) -> np.ndarray[K,d]
-        # Optional concise rule text to keep re-ranker context compact
-        self.rerank_rule_overrides: dict[str, str] = {}
-        self._last_diagnostics: dict[tuple[str, str], dict] = {}
-
-    def set_last_diagnostics(
-        self, unit_id: str, label_id: str, diagnostics: dict, original_unit_id: str | None = None
-    ):
-        """Cache the most recent retrieval diagnostics for a (unit, label) pair."""
-
-        key = (str(unit_id), str(label_id))
-        self._last_diagnostics[key] = diagnostics or {}
-        if original_unit_id is not None:
-            self._last_diagnostics[(str(original_unit_id), str(label_id))] = diagnostics or {}
-
-    def get_last_diagnostics(self, unit_id: str, label_id: str) -> dict:
-        return self._last_diagnostics.get((str(unit_id), str(label_id)), {})
-
-    def set_label_exemplars(self, label_id: str, rules: str, K: int, texts: list[str]):
-        key = (str(label_id), _stable_rules_hash(label_id, rules, K, getattr(self.models.embedder, "name_or_path", "")), int(K))
-        self._label_query_texts[key] = [t for t in texts if isinstance(t, str) and t.strip()]
-        if self._label_query_texts[key]:
-            E = self.store._embed(self._label_query_texts[key])   # (K, d)
-            self._label_query_embs[key] = E
-
-    def _get_label_query_embs(self, label_id: str, rules: str, K: int):
-        key = (str(label_id), _stable_rules_hash(label_id, rules, K, getattr(self.models.embedder, "name_or_path", "")), int(K))
-        return self._label_query_embs.get(key)
-
-    def _get_label_query_texts(self, label_id: str, rules: str, K: int):
-        key = (str(label_id), _stable_rules_hash(label_id, rules, K, getattr(self.models.embedder, "name_or_path", "")), int(K))
-        return self._label_query_texts.get(key)
-
-    def _rerank_rules_text(self, label_id: str, label_rules: str | None) -> str:
-        """Return a compact rule string when available for re-ranking only."""
-
-        if self.rerank_rule_overrides and label_id in self.rerank_rule_overrides:
-            return self.rerank_rule_overrides[label_id]
-        key = str(label_id)
-        if self.rerank_rule_overrides and key in self.rerank_rule_overrides:
-            return self.rerank_rule_overrides[key]
-        return label_rules or ""
-    
-    def _extract_meta(self, m:dict) -> dict:
-        """
-        Normalize chunk-level metadata into a compact dict that is JSON-safe and
-        useful for the LLM context header: {date, note_type, modality, title}.
-        Accepts raw chunk_meta dict.
-        """
-        out = {}
-        # Common fields
-        if "date_note" in m and m["date_note"]:
-            out["date"] = str(m["date_note"])
-
-        note_type = ""
-        if "notetype" in m and m["notetype"]:
-            note_type = str(m["notetype"])
-
-        # Try to parse richer metadata JSON if present (may include notetype)
-        meta_raw = m.get("document_metadata_json") or m.get("metadata_json")
-        meta_obj = _maybe_parse_jsonish(meta_raw)
-        if not note_type and isinstance(meta_obj, dict):
-            for key in ("note_type", "notetype", "noteType"):
-                val = meta_obj.get(key)
-                if val:
-                    note_type = str(val)
-                    break
-
-        if note_type:
-            out["note_type"] = note_type
-
-        out['other_meta'] = str(meta_raw)
-        return out
-
-    def _rr_key(self, q: str, t: str) -> str:
-        return stable_hash_pair(q or '', t or '')
-    
-
-    def _cross_scores_cached(self, query: str, texts: list) -> list:
-        missing_idx, miss_pairs, scores = [], [], [None] * len(texts)
-        for i, t in enumerate(texts):
-            k = self._rr_key(query, t)
-            sc = self._rr_cache.get(k)
-            if sc is None:
-                missing_idx.append(i); miss_pairs.append((query, t))
-            else:
-                scores[i] = sc
-        if missing_idx:
-            rr = self.models.reranker.predict(miss_pairs, batch_size=getattr(self.models, 'rerank_batch', 64), show_progress_bar=False)
-            for i, sc in zip(missing_idx, rr):
-                scores[i] = float(sc)
-                k = self._rr_key(query, texts[i])
-                # eviction handled by LRUCache automatically
-                self._rr_cache[k] = float(sc)
-        return scores
-
-    def _build_query(self, label_id: str, label_rules: Optional[str]) -> str:
-        if label_rules and isinstance(label_rules, str):
-            return label_rules.strip()
-        return ""
-
-    def _tokenize(self, text: str) -> List[str]:
-        if not text:
-            return []
-        return re.findall(r"\b\w+\b", str(text).lower())
-
-    def _tokenize_for_bm25(self, text: str) -> List[str]:
-        if hasattr(self.store, "_tokenize_for_bm25"):
-            try:
-                return self.store._tokenize_for_bm25(text)
-            except Exception:
-                pass
-        return self._tokenize(text)
-
-    def _bm25_index_for_patient(self, unit_id: str) -> Optional[dict]:
-        uid = str(unit_id)
-        cached = self._bm25_cache.get(uid)
-        if cached is not None:
-            return cached
-        if hasattr(self.store, "bm25_index_for_unit"):
-            idx = self.store.bm25_index_for_unit(uid)
-            if idx is not None:
-                self._bm25_cache[uid] = idx
-                return idx
-        idxs = self.store.get_patient_chunk_indices(uid)
-        if not idxs:
-            return None
-        docs, metas = [], []
-        for ix in idxs:
-            meta = self.store.chunk_meta[ix]
-            tokens = self._tokenize_for_bm25(meta.get("text", ""))
-            if not tokens:
-                continue
-            docs.append(tokens)
-            metas.append(meta)
-        if not docs:
-            return None
-        avgdl = sum(len(toks) for toks in docs) / float(len(docs))
-        index = {"docs": docs, "metas": metas, "avgdl": avgdl}
-        self._bm25_cache[uid] = index
-        if hasattr(self.store, "bm25_indices"):
-            try:
-                existing_indices = getattr(self.store, "bm25_indices", {})
-                if not isinstance(existing_indices, dict):
-                    existing_indices = {}
-                units = existing_indices.get("units") if "units" in existing_indices else existing_indices
-                if not isinstance(units, dict):
-                    units = {}
-                units = dict(units)
-                units[uid] = index
-                if "units" in existing_indices or not existing_indices:
-                    existing_indices = dict(existing_indices)
-                    existing_indices["units"] = units
-                else:
-                    existing_indices = units
-                self.store.bm25_indices = existing_indices
-                if hasattr(self.store, "_chunk_cache_dir_path"):
-                    self.store._save_bm25_indices(self.store._chunk_cache_dir_path, self.store.bm25_indices)
-            except Exception:
-                pass
-        return index
-
-    def _bm25_hits_for_patient(self, unit_id: str, keywords: List[str]) -> List[dict]:
-        if not keywords:
-            return []
-        index = self._bm25_index_for_patient(unit_id)
-        if not index:
-            return []
-        query_tokens = []
-        for kw in keywords:
-            if isinstance(kw, str):
-                query_tokens.extend(self._tokenize_for_bm25(kw))
-        if not query_tokens:
-            return []
-        docs = index["docs"]
-        idf = getattr(self.store, "idf_global", {}) or {}
-        if not idf and isinstance(getattr(self.store, "bm25_indices", None), dict):
-            indices_dict = getattr(self.store, "bm25_indices", {})
-            idf = indices_dict.get("idf_global", {}) if isinstance(indices_dict, dict) else {}
-        if not idf and hasattr(self.store, "_build_bm25_indices"):
-            built = self.store._build_bm25_indices()
-            if isinstance(built, dict):
-                self.store.bm25_indices = built
-                idf = built.get("idf_global", {})
-                if hasattr(self.store, "_chunk_cache_dir_path"):
-                    try:
-                        self.store._save_bm25_indices(self.store._chunk_cache_dir_path, built)
-                    except Exception:
-                        pass
-        if not idf:
-            return []
-        avgdl = index["avgdl"] or 1.0
-        metas = index["metas"]
-        k1, b = 1.5, 0.75
-        scores: List[tuple[float, dict]] = []
-        for toks, meta in zip(docs, metas):
-            tf = Counter(toks)
-            dl = len(toks)
-            score = 0.0
-            for term in query_tokens:
-                if term not in tf or term not in idf:
-                    continue
-                freq = tf[term]
-                denom = freq + k1 * (1 - b + b * dl / avgdl)
-                score += idf[term] * (freq * (k1 + 1)) / (denom + 1e-12)
-            if score <= 0:
-                continue
-            scores.append((score, meta))
-        scores.sort(key=lambda pair: pair[0], reverse=True)
-        out = []
-        for score, meta in scores[: self.cfg.keyword_topk]:
-            out.append({
-                "doc_id": meta.get("doc_id"),
-                "chunk_id": meta.get("chunk_id"),
-                "metadata": self._extract_meta(meta),
-                "text": meta.get("text", ""),
-                "score": float(score),
-                "source": "bm25",
-            })
-        return out
-
-    def _reciprocal_rank_fusion(self, runs: List[List[dict]], constant: int = 60) -> List[dict]:
-        fused: Dict[Tuple[str, int], Dict[str, Any]] = {}
-        for run in runs:
-            if not run:
-                continue
-            for rank, item in enumerate(run):
-                doc_id = str(item.get("doc_id"))
-                try:
-                    chunk_id = int(item.get("chunk_id", -1))
-                except Exception:
-                    continue
-                if chunk_id < 0:
-                    continue
-                key = (doc_id, chunk_id)
-                entry = fused.setdefault(key, {"score": 0.0, "item": item})
-                entry["score"] += 1.0 / (constant + rank + 1)
-                if entry["item"] is not item:
-                    merged = dict(entry["item"])
-                    for k, v in item.items():
-                        if k not in merged or merged[k] in (None, ""):
-                            merged[k] = v
-                    entry["item"] = merged
-        fused_items = []
-        for data in fused.values():
-            merged_item = dict(data.get("item", {}))
-            merged_item["score"] = float(data.get("score", 0.0))
-            fused_items.append(merged_item)
-        fused_items.sort(key=lambda d: d.get("score", 0.0), reverse=True)
-        return fused_items
-
-    def _mmr_select(self, q_emb: np.ndarray, cand_idxs: List[int], k: int, lam: float) -> List[int]:
-        if k<=0 or not cand_idxs: return []
-        X = self.store.X[cand_idxs]
-        Xn = X / (np.linalg.norm(X,axis=1,keepdims=True)+1e-12)
-        q = q_emb / (np.linalg.norm(q_emb)+1e-12)
-        simq = Xn @ q
-        sel = []; used = np.zeros(len(cand_idxs), dtype=bool)
-        for _ in range(min(k,len(cand_idxs))):
-            if not sel:
-                i = int(np.argmax(simq)); sel.append(i); used[i]=True; continue
-            M = Xn[[i for i,u in enumerate(used) if u]]
-            maxsim = (Xn @ M.T).max(axis=1)
-            scores = lam*simq - (1-lam)*maxsim
-            scores[used] = -1e9
-            i = int(np.argmax(scores)); sel.append(i); used[i]=True
-        return [cand_idxs[i] for i in sel]
-
-    def _mmr_select_ranked(
-        self,
-        cand_idxs: list[int],
-        rel_scores: list[float],
-        k: int,
-        lam: float,
-    ) -> list[int]:
-        """
-        Greedy MMR using CE relevance scores and embedding cosine similarity.
-        Expects cand_idxs to align 1:1 with rel_scores.
-        """
-        if k <= 0 or not cand_idxs:
-            return []
-        import numpy as _np
-
-        rel = _np.asarray(rel_scores, dtype=float)
-        X = self.store.X[cand_idxs]
-        Xn = X / (_np.linalg.norm(X, axis=1, keepdims=True) + 1e-12)
-        used = _np.zeros(len(cand_idxs), dtype=bool)
-        sel: list[int] = []
-        for _ in range(min(k, len(cand_idxs))):
-            if not sel:
-                i = int(_np.argmax(rel))
-            else:
-                M = Xn[used]
-                maxsim = (Xn @ M.T).max(axis=1)
-                scores = lam * rel - (1 - lam) * maxsim
-                scores[used] = -1e9
-                i = int(_np.argmax(scores))
-            sel.append(i)
-            used[i] = True
-        return [cand_idxs[i] for i in sel]
-
-    def _dedup_rerank(self, query: str, items: List[dict], final_topk: int) -> List[dict]:
-        if not items: return []
-        seen = set(); dedup = []
-        for it in items:
-            key = (it.get("doc_id"), it.get("chunk_id"))
-            if key in seen: continue
-            seen.add(key); dedup.append(it)
-        texts = [it["text"] for it in dedup]
-        rr = self._cross_scores_cached(query, texts)
-        for it, s in zip(dedup, rr):
-            it["score"] = float(s) + float(it.get("score",0.0))
-        dedup.sort(key=lambda d: d["score"], reverse=True)
-        return dedup[:final_topk]
-
-    def retrieve_for_patient_label(
-        self,
-        unit_id: str | None,
-        label_id: str,
-        label_rules: str | None,
-        topk_override: int | None = None,
-        min_k_override: int | None = None,
-        mmr_lambda_override: float | None = None,
-        *,
-        original_unit_id: str | None = None,
-    ) -> list[dict]:
-        """
-        Patient-only RAG for (unit_id, label_id).
-        Returns a list of snippets: [{"doc_id","chunk_id","text","score","source"}, ...]
-        No global/corpus search lives here; use expand_from_snippets for discovery across patients.
-
-        If mmr_lambda_override (or cfg.mmr_lambda) is set in [0,1], apply MMR on the CE-ranked pool
-        to pick a diverse final top-K.
-        """
-        import numpy as _np
-        
-        # ---- helpers ----
-        diagnostics: dict[str, object] = {
-            "unit_id": str(unit_id),
-            "label_id": str(label_id),
-            "stage": "init",
-        }
-
-        def _score_stats(items: list[dict]) -> dict:
-            scores = [float(it.get("score", 0.0)) for it in items if isinstance(it.get("score"), (int, float))]
-            if not scores:
-                return {}
-            return {
-                "min": float(min(scores)),
-                "max": float(max(scores)),
-                "mean": float(_np.mean(scores)),
-            }
-
-        def _patient_local_rank(_unit: str, _q_emb: _np.ndarray, need: int) -> list[dict]:
-            idxs = self.store.get_patient_chunk_indices(_unit)
-            if not idxs:
-                return []
-            X = self.store.X[idxs]
-            Xn = X / (_np.linalg.norm(X, axis=1, keepdims=True) + 1e-12)
-            qn = _q_emb / (_np.linalg.norm(_q_emb) + 1e-12)
-            sims = (Xn @ qn)
-            order = _np.argsort(-sims)[: max(need, 50)]
-            out = []
-            for j in order:
-                m = self.store.chunk_meta[idxs[j]]
-                out.append({
-                    "doc_id": m["doc_id"],
-                    "chunk_id": m["chunk_id"],
-                    "metadata": self._extract_meta(m),
-                    "text": m["text"],
-                    "score": float(sims[j]),
-                    "source": "patient_local",
-                })
-            return out
-        
-        def _patient_local_rank_multi(self, unit_id: str, Q: np.ndarray, need: int) -> list[dict]:
-            idxs = self.store.get_patient_chunk_indices(str(unit_id))
-            if not idxs:
-                return []
-            X  = self.store.X[idxs].astype("float32", copy=False)
-            Xn = X / (np.linalg.norm(X, axis=1, keepdims=True) + 1e-12)
-            Qn = Q / (np.linalg.norm(Q, axis=1, keepdims=True) + 1e-12)        # (K,d)
-            sims = Xn @ Qn.T                                                   # (n_chunks, K)
-            best = sims.max(axis=1)                                            # (n_chunks,)
-            order = np.argsort(-best)[: max(need, 50)]
-            out = []
-            for j in order:
-                m = self.store.chunk_meta[idxs[j]]
-                out.append({
-                    "doc_id": m["doc_id"],
-                    "chunk_id": m["chunk_id"],
-                    "text": m["text"],
-                    "score": float(best[j]),
-                    "source": "patient_local_multi"
-                })
-            return out
-        
-        def _neighbors(items: list[dict], hops: int = 1) -> list[dict]:
-            if hops <= 0:
-                return []
-            by_doc: dict[str, dict[int, int]] = {}
-            for ix, m in enumerate(self.store.chunk_meta):
-                by_doc.setdefault(m["doc_id"], {})[int(m["chunk_id"])] = ix
-        
-            out, seen = [], set((it["doc_id"], int(it["chunk_id"])) for it in items)
-            for it in items:
-                did, cid = it["doc_id"], int(it["chunk_id"])
-                table = by_doc.get(did, {})
-                for d in range(1, hops + 1):
-                    for cid2 in (cid - d, cid + d):
-                        ix = table.get(cid2)
-                        if ix is None:
-                            continue
-                        key = (did, cid2)
-                        if key in seen:
-                            continue
-                        seen.add(key)
-                        m = self.store.chunk_meta[ix]
-                        out.append({
-                            "doc_id": m["doc_id"],
-                            "chunk_id": m["chunk_id"],
-                            "metadata": self._extract_meta(m),
-                            "text": m["text"],
-                            "score": 0.0,
-                            "source": "neighbor",
-                        })
-            return out
-        
-        def _dedup_only(items: list[dict]) -> list[dict]:
-            if not items:
-                return []
-            seen, dedup = set(), []
-            for it in items:
-                key = (it.get("doc_id"), int(it.get("chunk_id")))
-                if key in seen:
-                    continue
-                seen.add(key)
-                dedup.append(it)
-            return dedup
-        
-        # ---- config ----
-        cfg_rag = getattr(self, "cfg", getattr(self, "rag", None)) or self.cfg
-        cfg_final_k = getattr(cfg_rag, "top_k_final", None)
-        final_k_raw = topk_override or cfg_final_k or getattr(cfg_rag, "per_label_topk", 6)
-        try:
-            final_k = max(1, int(final_k_raw))
-        except Exception:
-            final_k = 1
-        min_k     = min_k_override or max(1, getattr(cfg_rag, "min_context_chunks", 3))
-        mmr_mult  = max(1, getattr(cfg_rag, "mmr_multiplier", 3))  # pool size before CE = final_k * mmr_mult
-        hops      = int(getattr(cfg_rag, "neighbor_hops", 1))
-        keyword_fraction = float(getattr(cfg_rag, "keyword_fraction", 0.0))
-        keyword_fraction = max(0.0, min(1.0, keyword_fraction))
-        use_kw    = keyword_fraction > 0.0 and bool(getattr(cfg_rag, "use_keywords", True))
-        mmr_select_k = final_k * mmr_mult
-
-        # λ (0..1)
-        lam = mmr_lambda_override
-        if lam is None: lam = getattr(cfg_rag, "mmr_lambda", None)
-        lam = None if lam is None else float(lam)
-        if lam is not None: lam = max(0.0, min(1.0, lam))
-        diagnostics.update(
-            {
-                "stage": "config",
-                "rag_mode": "patient_local",
-                "final_k": final_k,
-                "min_k": min_k,
-                "mmr": {"lambda": lam, "multiplier": mmr_mult, "select_k": mmr_select_k},
-            }
-        )
-
-        # ---- label-aware query + embedding ----
-        try:
-            label_types = repo.label_types()
-        except Exception:
-            label_types = {}
-        base_k = getattr(self.cfg, "exemplar_K", None)
-        K_use = int(base_k if base_k is not None else 6)
-        if K_use <= 0:
-            K_use = 6
-        opts = _options_for_label(
-            label_id,
-            label_types.get(str(label_id), "categorical"),
-            getattr(self, "label_configs", {}),
-        ) or []
-        if opts:
-            K_use = max(K_use, len(opts))
-        cached_exemplar_embs = self._get_label_query_embs(label_id, label_rules, K=K_use)
-        exemplar_texts = self._get_label_query_texts(label_id, label_rules, K=K_use) or []
-
-        lblcfg = self.label_configs.get(label_id, {}) if isinstance(self.label_configs, dict) else {}
-        manual_query = None
-        if isinstance(lblcfg, Mapping):
-            raw_manual = lblcfg.get("search_query") or lblcfg.get("rag_query")
-            if isinstance(raw_manual, str) and raw_manual.strip():
-                manual_query = raw_manual.strip()
-
-        mmr_query_embs: list[np.ndarray] = []
-        semantic_queries_struct: list[SemanticQuery] = []
-
-        valid_exemplars = [t for t in exemplar_texts if isinstance(t, str) and t.strip()]
-
-        # Build prioritized query list: manual override > exemplars > label rules
-        if manual_query:
-            query_texts = [manual_query]
-            query_embs = [None]
-            query_sources = ["manual"]
-        elif valid_exemplars:
-            query_texts = list(valid_exemplars)
-            query_sources = ["exemplar"] * len(valid_exemplars)
-            if cached_exemplar_embs is not None and getattr(cached_exemplar_embs, "ndim", 1) == 2:
-                query_embs = [
-                    cached_exemplar_embs[i]
-                    for i in range(min(len(valid_exemplars), cached_exemplar_embs.shape[0]))
-                ]
-                if len(query_embs) < len(query_texts):
-                    # pad for later embedding
-                    query_embs.extend([None] * (len(query_texts) - len(query_embs)))
-            else:
-                query_embs = [None] * len(valid_exemplars)
-        else:
-            fallback_rules = (label_rules or "").strip()
-            query_texts = [fallback_rules]
-            query_embs = [None]
-            query_sources = ["rules"]
-
-        # Fill any missing embeddings
-        missing_idxs = [i for i, emb in enumerate(query_embs) if emb is None]
-        if missing_idxs:
-            missing_embs = list(self.store._embed([query_texts[i] for i in missing_idxs]))
-            for idx, emb in zip(missing_idxs, missing_embs):
-                query_embs[idx] = emb
-
-        query_source = query_sources[0] if query_sources else "rules"
-        rerank_query_texts = list(query_texts)
-
-        diagnostics.update(
-            {
-                "stage": "queries",
-                "manual_query": manual_query,
-                "queries": query_texts,
-                "exemplar_queries": exemplar_texts,
-                "query_source": query_source,
-                "query_sources": query_sources,
-                "rerank_queries": rerank_query_texts,
-            }
-        )
-
-        for q_text, q_emb, q_src in zip(query_texts, query_embs, query_sources):
-            semantic_queries_struct.append(
-                SemanticQuery(text=q_text, embedding=q_emb, source=q_src)
-            )
-            if q_emb is not None:
-                mmr_query_embs.append(q_emb)
-
-        q_emb = np.mean(np.vstack(mmr_query_embs), axis=0) if mmr_query_embs else None
-        query = query_texts[0] if query_texts else ""
-
-        keywords: list[str] = []
-        if use_kw:
-            lblcfg = self.label_configs.get(label_id, {}) if isinstance(self.label_configs, dict) else {}
-            cfg_keywords = getattr(cfg_rag, "keywords", [])
-            if isinstance(cfg_keywords, (list, tuple)):
-                keywords.extend(str(k) for k in cfg_keywords if isinstance(k, str) and k.strip())
-            per_label_kw_cfg = getattr(cfg_rag, "label_keywords", {})
-            if isinstance(per_label_kw_cfg, Mapping):
-                label_kw = per_label_kw_cfg.get(label_id) or per_label_kw_cfg.get(str(label_id))
-                if isinstance(label_kw, (list, tuple)):
-                    keywords.extend(str(k) for k in label_kw if isinstance(k, str) and k.strip())
-            lbl_keywords = lblcfg.get("keywords", []) if isinstance(lblcfg, Mapping) else []
-            if isinstance(lbl_keywords, (list, tuple)):
-                keywords.extend(str(k) for k in lbl_keywords if isinstance(k, str) and k.strip())
-
-        uniq_keywords: list[str] = []
-        if keywords:
-            seen_kw = set()
-            for kw in keywords:
-                if kw in seen_kw:
-                    continue
-                seen_kw.add(kw)
-                uniq_keywords.append(kw)
-
-        pool_factor = int(getattr(cfg_rag, "pool_factor", 3))
-        pool_oversample = float(getattr(cfg_rag, "pool_oversample", 1.5))
-
-        coordinator = RetrievalCoordinator(
-            semantic_searcher=lambda sq, need: _patient_local_rank(str(unit_id), sq.embedding, need),
-            keyword_searcher=lambda kws, need: (self._bm25_hits_for_patient(str(unit_id), kws) or [])[:need],
-        )
-
-        pool, pool_diag = coordinator.build_candidate_pool(
-            semantic_queries=semantic_queries_struct,
-            keywords=uniq_keywords,
-            top_k_final=final_k,
-            keyword_fraction=keyword_fraction,
-            pool_factor=pool_factor,
-            oversample=pool_oversample,
-        )
-
-        diagnostics["retrieval_coordinator"] = pool_diag
-        diagnostics["keyword_search"] = {
-            "keywords": uniq_keywords,
-            "hit_count": pool_diag.get("keyword_hits", 0),
-        }
-        diagnostics["semantic_search"] = {
-            "runs": len(semantic_queries_struct),
-            "score_stats": _score_stats(pool),
-        }
-        diagnostics["pool"] = {
-            "semantic": pool_diag.get("semantic_hits", 0),
-            "bm25": pool_diag.get("keyword_hits", 0),
-            "deduped": pool_diag.get("final_pool", 0),
-            "targets": {
-                "semantic": pool_diag.get("semantic_target", 0),
-                "keyword": pool_diag.get("keyword_target", 0),
-                "pool": pool_diag.get("pool_target", 0),
-            },
-        }
-
-        # Neighbor padding for short pools
-        if len(pool) < max(min_k, final_k):
-            extra = _neighbors(pool, hops=hops)
-            pool = _dedup_only(pool + extra)
-            diagnostics["pool"]["neighbors_added"] = len(pool) - diagnostics["pool"].get("deduped", 0)
-        if not pool:
-            diagnostics["stage"] = "empty_pool"
-            self.set_last_diagnostics(unit_id, label_id, diagnostics, original_unit_id=original_unit_id)
-            return []
-    
-        # Build (doc,chunk) -> ix map with STRING doc_id, INT chunk_id
-        by_doc: dict[str, dict[int, int]] = {}
-        for ix, m in enumerate(self.store.chunk_meta):
-            by_doc.setdefault(str(m["doc_id"]), {})[int(m["chunk_id"])] = ix
-    
-        cand_idxs, cand_items = [], []
-        for it in pool:
-            did = str(it.get("doc_id")); cid = int(it.get("chunk_id"))
-            ix = by_doc.get(did, {}).get(cid)
-            if ix is not None:
-                cand_idxs.append(ix); cand_items.append(it)
-
-        diagnostics.setdefault("pool", {})["source_counts"] = dict(
-            Counter(str(it.get("source")) for it in pool)
-        )
-
-        if q_emb is None:
-            embed_text = query if query_texts else ""
-            q_emb = self.store._embed([embed_text])[0]
-
-        def _cross_scores_for_queries(q_texts: list[str], cand_texts: list[str]) -> list[float]:
-            if not q_texts:
-                return [float(s) for s in self._cross_scores_cached(query, cand_texts)]
-            per_query = [self._cross_scores_cached(qt, cand_texts) for qt in q_texts]
-            return list(np.max(np.vstack(per_query), axis=0)) if per_query else [0.0] * len(cand_texts)
-
-        # CE fallback if mapping failed
-        if not cand_idxs:
-            texts = [it["text"] for it in pool]
-            rr = _cross_scores_for_queries(rerank_query_texts, texts)
-            for it, s in zip(pool, rr):
-                it["score"] = float(s)
-            pool.sort(key=lambda d: d["score"], reverse=True)
-            return pool[:final_k]
-    
-        # Preselect for CE scoring (head)
-        k_pre = min(len(cand_items), max(final_k, min_k, mmr_select_k))
-        diagnostics.setdefault("mmr", {}).update(
-            {"candidate_pool": len(cand_items), "select_size": k_pre}
-        )
-        pre: list[dict] = []
-        pre_idxs: list[int] = []
-        for ix, it in zip(cand_idxs[:k_pre], cand_items[:k_pre]):
-            pre.append(it)
-            pre_idxs.append(ix)
-
-        # CE last, CE-only score
-        texts = [it["text"] for it in pre]
-        rr = _cross_scores_for_queries(rerank_query_texts, texts)
-        for it, s in zip(pre, rr):
-            it["score"] = float(s)
-
-        # Sort by CE to feed MMR or take head
-        scored = [
-            {"item": it, "store_idx": ix}
-            for it, ix in sorted(
-                zip(pre, pre_idxs), key=lambda t: t[0].get("score", 0.0), reverse=True
-            )
-        ]
-
-        if lam is not None:
-            sel = self._mmr_select_ranked(
-                [d["store_idx"] for d in scored],
-                [float(d["item"].get("score", 0.0)) for d in scored],
-                k=final_k,
-                lam=lam,
-            )
-            idx_to_item = {d["store_idx"]: d["item"] for d in scored}
-            out = [idx_to_item[i] for i in sel if i in idx_to_item]
-        else:
-            out = [d["item"] for d in scored[:final_k]]
-
-        diagnostics.setdefault("mmr", {}).update(
-            {
-                "used_lambda": lam,
-                "pre_ce_count": len(pre),
-                "final_before_topoff": len(out),
-                "pre_score_stats": _score_stats(pre),
-            }
-        )
-
-        # Top-off to min_k using CE on the rest
-        if len(out) < min_k:
-            picked = set((o["doc_id"], int(o["chunk_id"])) for o in out)
-            rest = [it for it in pool if (it["doc_id"], int(it["chunk_id"])) not in picked]
-            if rest:
-                rr2 = _cross_scores_for_queries(rerank_query_texts, [it["text"] for it in rest])
-                for it, s in zip(rest, rr2):
-                    it["score"] = float(s)
-                rest.sort(key=lambda d: d["score"], reverse=True)
-                need = min_k - len(out)
-                out.extend(rest[:need])
-
-        diagnostics["final_selection"] = {
-            "count": len(out),
-            "score_stats": _score_stats(out),
-        }
-
-        diagnostics["stage"] = "complete"
-        self.set_last_diagnostics(unit_id, label_id, diagnostics, original_unit_id=original_unit_id)
-        return out[:final_k]
-
-    def expand_from_snippets(self, label_id: str, snippets: List[str], seen_pairs: set, per_seed_k: int=100) -> Dict[str,float]:
-        out: Dict[str,float] = {}
-        for sn in snippets:
-            if not sn: continue
-            sims, idxs = self.store.search([sn], topk=max(per_seed_k, 50))
-            idxs = idxs[0]; sims = sims[0]
-            pre = []
-            for sc, ix in zip(sims, idxs):
-                if ix < 0: continue
-                m = self.store.chunk_meta[ix]
-                pre.append((ix, m["text"], m["unit_id"], float(sc)))
-            texts = [txt for _, txt, _, _ in pre]
-            rr = self._cross_scores_cached(sn, texts)
-            for (ix, _txt, uid, _sc), s in zip(pre, rr):
-                if (uid, label_id) in seen_pairs: continue
-                out[uid] = max(out.get(uid, 0.0), float(s))
-        return out
-
 
 # ------------------------------
 # Label-aware pooling with prototypes
@@ -1243,931 +315,6 @@ def kcenter_select(vecs: np.ndarray, k: int, seed_idx: Optional[int]=None, prese
 # ------------------------------
 # LLM-first probe + forced-choice + expansion
 # ------------------------------
-def _options_for_label(label_id: str, label_type: str, label_cfgs: dict) -> Optional[List[str]]:
-    cfg = label_cfgs.get(label_id, {}) if isinstance(label_cfgs, dict) else {}
-    opts = cfg.get("options")
-    if isinstance(opts, list) and 2 <= len(opts) <= 26:
-        return [str(o) for o in opts]
-    return None
-
-# ------------------------------
-# Selection helpers
-# ------------------------------
-# Parent-Child Gating + Family Labeling
-# ------------------------------
-
-def _parse_date(x):
-    try:
-        import pandas as _pd
-        if x is None or (isinstance(x, float) and math.isnan(x)): return None
-        dt = _pd.to_datetime(x, errors="coerce")
-        if dt is _pd.NaT: return None
-        return dt
-    except Exception:
-        return None
-
-def _to_number(x):
-    try:
-        if isinstance(x, (int, float)) and not (isinstance(x, float) and math.isnan(x)):
-            return float(x)
-        s = str(x).strip()
-        s = s.replace(',', '')
-        return float(s)
-    except Exception:
-        return None
-
-def _canon_str(x):
-    if x is None:
-        return ""
-    s = str(x)
-    s = re.sub(r"\s+", " ", s).strip()
-    return s
-
-def _canon_cat(x):
-    s = _canon_str(x).lower()
-    # normalize common truthy/falsey
-    if s in {"y","yes","true","1","present","positive","pos"}:
-        return "yes"
-    if s in {"n","no","false","0","absent","negative","neg"}:
-        return "no"
-    return s
-
-def build_label_dependencies(label_config: dict) -> tuple[dict, dict, list]:
-    """Return (parent->children, child->parents, roots) from label_config.
-    Supports either:
-      label_config[<lid>]['gated_by'] = <parent or [parents]>
-      label_config[<lid>]['gating_rules'] = list of rule dicts with keys:
-          - parent: parent label_id (optional if 'gated_by' provided)
-          - type: 'categorical'|'numeric'|'date'
-          - op: one of 'in','notin','==','!=','>','>=','<','<=','between','outside','exists','notnull'
-          - values/value/low/high/inclusive (depending on op)
-      label_config[<lid>]['gating_logic'] = 'AND'|'OR' (default AND) across provided rules
-    """
-    parent_to_children: dict[str, list[str]] = {}
-    child_to_parents: dict[str, list[str]] = {}
-    if not isinstance(label_config, dict):
-        return {}, {}, []
-    for lid, cfg in label_config.items():
-        if str(lid) == "_meta":
-            continue
-        if not isinstance(cfg, dict):
-            continue
-        gb = cfg.get('gated_by')
-        # Normalize to list
-        parents = []
-        if gb:
-            if isinstance(gb, (list, tuple, set)):
-                parents.extend([str(x) for x in gb])
-            else:
-                parents.append(str(gb))
-        # Also scan rules to pick parent fields
-        rules = cfg.get('gating_rules') or []
-        for r in (rules if isinstance(rules, list) else [rules]):
-            p = None
-            if isinstance(r, dict):
-                p = r.get('parent') or r.get('gated_by') or r.get('field')
-            if p:
-                p = str(p)
-                if p not in parents:
-                    parents.append(p)
-        if parents:
-            for p in parents:
-                parent_to_children.setdefault(str(p), []).append(str(lid))
-                child_to_parents.setdefault(str(lid), []).append(str(p))
-    all_labels = {str(k) for k in label_config.keys() if str(k) != "_meta"}
-    roots = [lid for lid in all_labels if lid not in child_to_parents]
-    return parent_to_children, child_to_parents, roots
-
-def _check_rule(parent_value, parent_type: str, rule: dict) -> bool:
-    """Evaluate a single gating rule for a parent value of a given type."""
-    if parent_type == 'numeric':
-        v = _to_number(parent_value)
-        if v is None: 
-            return False
-        op = str(rule.get('op','in')).lower()
-        if op == 'between':
-            lo = _to_number(rule.get('low', None))
-            hi = _to_number(rule.get('high', None))
-            inc = bool(rule.get('inclusive', True))
-            if lo is None or hi is None: 
-                return False
-            return (lo <= v <= hi) if inc else (lo < v < hi)
-        elif op in ('>','>=','<','<=','==','!='):
-            val = _to_number(rule.get('value', rule.get('values', [None])[0] if isinstance(rule.get('values'), list) else None))
-            if val is None: 
-                return False
-            if op == '>':  return v >  val
-            if op == '>=': return v >= val
-            if op == '<':  return v <  val
-            if op == '<=': return v <= val
-            if op == '==': return v == val
-            if op == '!=': return v != val
-            return False
-        elif op in ('in','notin'):
-            vals = rule.get('values', [])
-            vals = [ _to_number(x) for x in (vals if isinstance(vals, list) else [vals]) ]
-            vals = [x for x in vals if x is not None]
-            ok = v in vals
-            return ok if op == 'in' else (not ok)
-        elif op in ('exists','notnull'):
-            return v is not None
-        else:
-            return False
-    elif parent_type == 'date':
-        d = _parse_date(parent_value)
-        if d is None: 
-            return False
-        op = str(rule.get('op','>')).lower()
-        if op == 'between':
-            lo = _parse_date(rule.get('low'))
-            hi = _parse_date(rule.get('high'))
-            inc = bool(rule.get('inclusive', True))
-            if lo is None or hi is None: 
-                return False
-            return (lo <= d <= hi) if inc else (lo < d < hi)
-        elif op in ('>','>=','<','<=','==','!='):
-            val = _parse_date(rule.get('value', rule.get('values', [None])[0] if isinstance(rule.get('values'), list) else None))
-            if val is None: 
-                return False
-            if op == '>':  return d >  val
-            if op == '>=': return d >= val
-            if op == '<':  return d <  val
-            if op == '<=': return d <= val
-            if op == '==': return d == val
-            if op == '!=': return d != val
-            return False
-        elif op in ('exists','notnull'):
-            return d is not None
-        else:
-            return False
-    else:
-        s = _canon_cat(parent_value)
-        op = str(rule.get('op','in')).lower()
-        if op in ('in','notin'):
-            vals = rule.get('values', [])
-            vals = [ _canon_cat(x) for x in (vals if isinstance(vals, list) else [vals]) ]
-            ok = s in vals
-            return ok if op == 'in' else (not ok)
-        elif op in ('==','!='):
-            val = _canon_cat(rule.get('value', rule.get('values', [None])[0] if isinstance(rule.get('values'), list) else None))
-            if op == '==': return s == val
-            else: return s != val
-        elif op in ('exists','notnull'):
-            return len(s) > 0
-        else:
-            # default: treat as truthy yes
-            return s in {'yes','present','true','1'}
-
-def _gating_for_child(child_id: str, label_config: dict) -> dict:
-    cfg = (label_config or {}).get(child_id, {}) if isinstance(label_config, dict) else {}
-    rules = cfg.get('gating_rules') or []
-    if isinstance(rules, dict):
-        rules = [rules]
-    logic = str(cfg.get('gating_logic','AND')).upper()
-    parents_declared = cfg.get('gated_by')
-    parents = []
-    if parents_declared:
-        parents = parents_declared if isinstance(parents_declared, list) else [parents_declared]
-        # If no explicit rules, build a default categorical truthy rule for each parent
-        if not rules:
-            rules = [ {'parent': p, 'type': 'categorical', 'op': 'in', 'values': ['yes','present','true','1']} for p in parents ]
-    # ensure parent field filled in rules
-    for r in rules:
-        if isinstance(r, dict) and 'parent' not in r:
-            if parents:
-                r['parent'] = parents[0]
-    return {'rules': rules, 'logic': logic}
-
-def evaluate_gating(child_id: str, unit_id: str, parent_preds: dict, label_types: dict, label_config: dict) -> bool:
-    """Return True if child is eligible given current parent predictions."""
-    g = _gating_for_child(child_id, label_config)
-    rules = g.get('rules') or []
-    if not rules:
-        return True
-    logic = g.get('logic','AND').upper()
-    outcomes = []
-    for r in rules:
-        if not isinstance(r, dict): 
-            continue
-        p = str(r.get('parent'))
-        p_type = label_types.get(p, 'categorical')
-        val = parent_preds.get((unit_id, p))
-        outcomes.append(_check_rule(val, p_type, r))
-    if not outcomes:
-        return True
-    return all(outcomes) if logic != 'OR' else any(outcomes)
-
-class FamilyLabeler:
-    """Label entire family tree per unit, honoring parent-child gating."""
-    def __init__(self, llm: LLMLabeler, retriever: RAGRetriever, repo: DataRepository, label_config: dict, scCfg: SCJitterConfig, llmfirst_cfg: LLMFirstConfig):
-        self.llm = llm
-        self.retriever = retriever
-        self.repo = repo
-        self.label_config = label_config or {}
-        self.scCfg = scCfg
-        self.cfg = llmfirst_cfg
-        self.context_builder = getattr(retriever, "context_builder", None)
-        self.parent_to_children, self.child_to_parents, self.roots = build_label_dependencies(self.label_config)
-        
-    
-    def ensure_label_exemplars(self, rules_map: dict[str, str], K: int = 6):
-        try:
-            label_types = self.repo.label_types()
-        except Exception:
-            label_types = {}
-        for lid, rules in (rules_map or {}).items():
-            base_k = getattr(self.cfg, "exemplar_K", None)
-            K_use = int(base_k if base_k is not None else K)
-            if K_use <= 0:
-                K_use = K
-            opts = _options_for_label(lid, label_types.get(str(lid), "categorical"), self.label_config) or []
-            if opts:
-                K_use = max(K_use, len(opts))
-            # skip if already cached
-            if self.retriever._get_label_query_embs(lid, rules, K_use) is not None:
-                continue
-            t = self._generate_label_exemplars(lid, rules, K_use)
-            if not t:
-                t = self._fallback_label_exemplars(lid, rules, K_use, opts)
-            if t:
-                self.retriever.set_label_exemplars(lid, rules, K_use, t)
-
-    def _fallback_label_exemplars(self, label_id: str, rules: str, K: int, options: list[str] | None) -> list[str]:
-        """Graceful fallback if exemplar generation fails (e.g., Azure JSON mode issues).
-
-        We synthesize minimal exemplars using the label rules so downstream RAG still
-        has deterministic label-aware queries instead of aborting with
-        "label_exemplar_error". These are intentionally simple and will be embedded
-        just like LLM-generated snippets.
-        """
-        K = max(1, int(K))
-        base = self.retriever._build_query(label_id, rules)
-        texts: list[str] = []
-        opts = options or []
-        if opts:
-            for o in opts:
-                texts.append(f"{base} Example of option '{o}'.")
-        if not texts:
-            texts.append(base)
-        while len(texts) < K:
-            texts.append(base)
-        return texts[:K]
-
-    def _generate_label_exemplars(
-        self,
-        label_id: str,
-        rules: str,
-        K: int,
-        options: list[str] | None = None,
-    ) -> list[str]:
-        """
-        Generate K realistic clinical snippets (1–3 sentences) that satisfy the rule.
-        - If the label has options (categorical/binary), request JSON objects with {"option","text"}
-          and return a mix across options when feasible.
-        - If no options (numeric/date/etc.), request JSON with {"snippets": [<string>, ...]}.
-        - Azure OpenAI JSON mode (json_object); robust parsing and de-dup; returns List[str].
-        """
-        import time as _time
-        import json as _json
-    
-        # Resolve options from label_config if not explicitly provided
-        opts = options
-        if not opts:
-            try:
-                lblcfgs = getattr(self.retriever, "label_configs", {}) or {}
-                cfg = lblcfgs.get(label_id, {}) if isinstance(lblcfgs, dict) else {}
-                raw = cfg.get("options")
-                if isinstance(raw, list) and 2 <= len(raw) <= 26:
-                    opts = [str(x) for x in raw]
-            except Exception:
-                opts = None
-        has_opts = bool(opts)
-    
-        # ----- Build prompt (conditional on options) -----
-        if has_opts:
-            menu_text = "\n".join(f"- {o}" for o in opts)
-            system = (
-                "You write realistic clinical note snippets.\n"
-                "Output policy:\n"
-                "- Return ONLY a valid JSON object with a single key 'snippets', whose value is an array of objects.\n"
-                "- Each object MUST have: {\"option\": <string>, \"text\": <string>}.\n"
-                "- No extra keys, no commentary."
-            )
-            user = f"""
-                Produce {K} diverse clinical note snippets (2-4 sentences each) that would *satisfy* this label rule.
-                
-                Label: {label_id}
-                Rule: {rules or '(no extra rules)'}
-                Options:
-                {menu_text}
-                
-                Constraints:
-                - Style: realistic, diverse, clinician-authored note sentences. No headers, no lists, no commentary.
-                - No meta text (no “the patient meets criteria because...”, no references to “label” or “rule”).
-                - Vary phrasing, timing (present/past/relative dates), and synonyms.
-                - Keep PHI generic.
-                
-                Provide a MIX across options so each option appears at least once if feasible.
-                Set "option" to exactly one of: {", ".join(opts)}.
-                
-                Return ONLY:
-                {{
-                  "snippets": [
-                    {{"option": "<one of the options>", "text": "<2–4 sentence snippet>"}}
-                  ]
-                }}
-                """.strip()
-        else:
-            system = (
-                "You write realistic clinical note snippets.\n"
-                "Output policy:\n"
-                "- Return ONLY a valid JSON object with a single key 'snippets', whose value is an array of strings.\n"
-                "- Do NOT include an 'option' field.\n"
-                "- No extra keys, no commentary."
-            )
-            user = f"""
-                Produce {K} diverse clinical note snippets (2–4 sentences each) that would *satisfy* this label rule.
-                
-                Label: {label_id}
-                Rule: {rules or '(no extra rules)'}
-                
-                Constraints:
-                - Style: realistic, diverse, clinician-authored note sentences. No headers, no lists, no commentary.
-                - No meta text (no “the patient meets criteria because...”, no references to “label” or “rule”).
-                - Vary phrasing, timing (present/past/relative dates), and synonyms.
-                - Keep PHI generic.
-                
-                Return ONLY:
-                {{
-                  "snippets": ["<2–4 sentence snippet>", "..."]
-                }}
-                """.strip()
-    
-        # ----- Backend JSON call -----
-        temp = float(getattr(self.cfg, "exemplar_temperature", 0.7) or 0.7)
-
-        if has_opts:
-            snippets_schema: Mapping[str, Any] = {
-                "type": "object",
-                "properties": {
-                    "snippets": {
-                        "type": "array",
-                        "items": {
-                            "type": "object",
-                            "properties": {
-                                "option": {"type": "string"},
-                                "text": {"type": "string"},
-                            },
-                            "required": ["option", "text"],
-                            "additionalProperties": False,
-                        },
-                    }
-                },
-                "required": ["snippets"],
-                "additionalProperties": False,
-            }
-        else:
-            snippets_schema = {
-                "type": "object",
-                "properties": {
-                    "snippets": {
-                        "type": "array",
-                        "items": {"type": "string"},
-                    }
-                },
-                "required": ["snippets"],
-                "additionalProperties": False,
-            }
-
-        try:
-            result = self.llm.backend.json_call(
-                messages=[{"role": "system", "content": system}, {"role": "user", "content": user}],
-                temperature=temp,
-                logprobs=False,
-                top_logprobs=None,
-                response_format={
-                    "type": "json_object",
-                    "json_schema": snippets_schema,
-                },
-            )
-            content = result.content
-            # Capture the exact prompt + minimal context identifiers used this vote
-            LLM_RECORDER.record("label_exemplar", {
-                "label_id": label_id,
-                "prompt": {"system": system, "user": user},
-                "params": {
-                    "temperature": temp,
-                },
-                "output": content,
-            })
-        except Exception:
-            LLM_RECORDER.record("label_exemplar_error", {})
-            return []
-        
-        # ----- Parse JSON -----
-        texts_by_opt: dict[str, list[str]] = {}
-        texts: list[str] = []
-    
-        def _clean_text(s):
-            return str(s).strip()
-    
-        def _coerce_option(o: str) -> str:
-            if not has_opts or not isinstance(o, str):
-                return str(o or "")
-            lut = {s.lower(): s for s in opts}
-            return lut.get(o.lower(), str(o).strip())
-    
-        try:
-            data = _json.loads(content) if isinstance(content, str) else content
-            if has_opts:
-                objs = (data or {}).get("snippets", []) if isinstance(data, dict) else []
-                for it in objs:
-                    if not isinstance(it, dict): 
-                        continue
-                    txt = _clean_text(it.get("text", ""))
-                    opt = _coerce_option(it.get("option"))
-                    if len(txt) > 20 and ((not has_opts) or (opt in (opts or []))):
-                        texts_by_opt.setdefault(opt, []).append(txt)
-            else:
-                arr = (data or {}).get("snippets", []) if isinstance(data, dict) else (data if isinstance(data, list) else [])
-                for s in arr:
-                    txt = _clean_text(s)
-                    if len(txt) > 20:
-                        texts.append(txt)
-        except Exception:
-            # Fallback: try to pull the first JSON array
-            try:
-                i = content.find("["); j = content.rfind("]")
-                arr = _json.loads(content[i:j+1]) if (i != -1 and j != -1 and j > i) else []
-                for s in arr:
-                    txt = _clean_text(s)
-                    if len(txt) > 20:
-                        texts.append(txt)
-            except Exception:
-                pass
-    
-        # ----- De-dup and balance / limit to K -----
-        if has_opts:
-            # De-dup within each option bucket while preserving order
-            for o in list(texts_by_opt.keys()):
-                seen = set(); ded = []
-                for t in texts_by_opt[o]:
-                    if t not in seen:
-                        seen.add(t); ded.append(t)
-                texts_by_opt[o] = ded
-    
-            # Round-robin across options to get a mix, then fill remaining
-            out: list[str] = []
-            # one per option if feasible
-            for o in opts:
-                if len(out) >= K: break
-                if texts_by_opt.get(o):
-                    out.append(texts_by_opt[o].pop(0))
-            # fill remaining round-robin
-            while len(out) < K:
-                progressed = False
-                for o in opts:
-                    if len(out) >= K: break
-                    bucket = texts_by_opt.get(o) or []
-                    if bucket:
-                        out.append(bucket.pop(0)); progressed = True
-                if not progressed:
-                    break
-            # if still short, append any leftover
-            if len(out) < K:
-                for o in opts:
-                    for t in texts_by_opt.get(o, []):
-                        if len(out) >= K: break
-                        out.append(t)
-                    if len(out) >= K: break
-            return out[:K]
-        else:
-            # De-dup and cap
-            seen = set(); out = []
-            for t in texts:
-                if t not in seen:
-                    seen.add(t); out.append(t)
-                if len(out) >= K:
-                    break
-            return out
-
-
-    def label_family_for_unit(self, unit_id: str, label_types: dict[str,str], per_label_rules: dict[str,str], *,
-                              json_only: bool=False, json_n_consistency: int=1, json_jitter: bool=False) -> list[dict]:
-        """Label all parents then gated children for a unit. Returns list of probe rows."""
-        self.ensure_label_exemplars(per_label_rules)
-        results = []
-        # We'll process in a BFS: queue starts with roots, then expand children once parent gating satisfied
-        # Maintain map of parent predictions for gating evaluation
-        parent_preds: dict[tuple[str,str], Any] = {}
-        # Determine an evaluation order by repeatedly scanning for labels whose parents are satisfied or none
-        # But to preserve minimal disruption, we simply try roots then others with gating check at point of execution.
-        all_labels = list({*self.roots, *[str(lid) for lid in label_types.keys()]})
-        # We'll ensure unique and stable order
-        seen = set()
-        order = []
-        for lid in self.roots + [l for l in all_labels if l not in self.roots]:
-            if lid not in seen:
-                seen.add(lid); order.append(lid)
-        for lid in order:
-            ltype = label_types.get(lid, 'categorical')
-            rules = per_label_rules.get(lid, "")
-            # if this label is gated (is a child), check gating
-            allowed = evaluate_gating(lid, unit_id, parent_preds, label_types, self.label_config)
-            if not allowed:
-                # record gated-out for transparency but do not include in probe_df to avoid selection noise
-                continue
-            label_rows = self.llm.label_unit(
-                unit_id,
-                [lid],
-                label_types=label_types,
-                per_label_rules=per_label_rules,
-                context_builder=self.context_builder,
-                retriever=self.retriever,
-                llmfirst_cfg=self.cfg,
-                json_only=json_only,
-                json_n_consistency=json_n_consistency,
-                json_jitter=json_jitter,
-            )
-            if not label_rows:
-                continue
-            row = label_rows[0]
-            parent_preds[(unit_id, lid)] = row.get("prediction")
-            results.append(row)
-        return results
-
-    def sample_units_for_probe_enriched(self, per_label_rules: dict[str, str], exclude_units: Optional[set[str]] = None,) -> List[str]:
-        """
-        CE-ranked enrichment that reuses retriever.retrieve_for_patient_label:
-          For each parent P:
-            - Randomly sample units from unseen pool
-            - Call retrieve_for_patient_label(unit, P, rule) to get top-k patient snippets (already CE-ranked)
-            - Unit score = max/mean CE score over top m snippets
-          Then allocate quotas across parents, resolve overlaps by best parent, mix with uniform, and top up.
-        """
-        import numpy as np
-    
-        rng = np.random.default_rng()
-        ex = set(exclude_units or [])
-        
-        #universe after exclusion
-        all_units = [str(u) for u in sorted(self.repo.notes["unit_id"].unique().tolist()) if str(u) not in ex]
-        if not all_units:
-            return []
-
-        # --- overall mix ---
-        # uniform slice only from non-excluded
-        n_total  = int(getattr(self.cfg, "n_probe_units", 100) or 100)
-        mix      = float(getattr(self.cfg, "probe_enrichment_mix", 0.90) or 0.90)
-        n_enr    = int(round(n_total * mix))
-        n_unif   = max(0, n_total - n_enr)
-        uniform_units = rng.choice(all_units, size=min(n_unif, len(all_units)), replace=False).tolist() if n_unif > 0 else []
-
-    
-        # --- parents & pool ---
-        parents = list(getattr(self, "roots", []) or list((per_label_rules or {}).keys()))
-        if not parents or not all_units:
-            # fall back to uniform if nothing sensible
-            print('falling back to uniform')
-            return list(np.random.default_rng().choice(all_units, size=min(n_total, len(all_units)), replace=False)) if all_units else []
-        parent_scores: dict[str, list[tuple[str, float]]] = {}
-        best_parent: dict[str, tuple[str, float]] = {}
-        
-        # --- knobs for CE enrichment ---
-        per_parent_sample = int(getattr(self.cfg, "probe_ce_unit_sample", 800) or 800)
-        equalize         = bool(getattr(self.cfg, "probe_enrichment_equalize", True))
-    
-        # ---- score units per parent using your retrieve_for_patient_label ----
-        for p in parents:
-            rule = (per_label_rules or {}).get(p) or p
-            # candidate pool excludes both uniform slice and excluded units
-            pool_space = [u for u in all_units if (u not in set(uniform_units))]
-            if not pool_space:
-                parent_scores[p] = []
-                continue
-    
-            per_parent_sample = int(getattr(self.cfg, "probe_ce_unit_sample", 800) or 800)
-            cand_units = rng.choice(pool_space, size=min(per_parent_sample, len(pool_space)), replace=False).tolist()
-            
-            _progress_every = float(self.cfg.progress_min_interval_s or 1.0)
-            unit_chunks, total_pairs = [], 0
-            
-            for uid in iter_with_bar(
-                    step = "Enriching probe pool",
-                    iterable = cand_units,
-                    total = len(cand_units),
-                    min_interval_s=_progress_every):
-                if uid in ex:             # redundant but explicit
-                    continue
-                ctx = self.context_builder.build_context_for_label(
-                    uid,
-                    p,
-                    rule,
-                    topk_override=int(getattr(self.cfg, "probe_ce_search_topk_per_unit", 24) or 24),
-                    single_doc_context_mode=getattr(self.cfg, "single_doc_context", "rag"),
-                    full_doc_char_limit=getattr(self.cfg, "single_doc_full_context_max_chars", None),
-                )
-                if not ctx:
-                    continue
-                unit_chunks.append((uid, ctx))
-                total_pairs += len(ctx)
-    
-            if not unit_chunks:
-                parent_scores[p] = []
-                continue
-    
-            # CE scores already in ctx['score']; aggregate top-m
-            m = int(getattr(self.cfg, "probe_ce_rerank_m", 3) or 3)
-            unit_rows = []
-            for uid, ctx in unit_chunks:
-                top = ctx[:max(1, m)]
-                vals = [float(it.get("score", 0.0)) for it in top]
-                unit_score = float(np.mean(vals)) if str(getattr(self.cfg, "probe_ce_unit_agg", "max")).lower() == "mean" else float(np.max(vals))
-                unit_rows.append((uid, unit_score))
-                prev = best_parent.get(uid)
-                if (prev is None) or (unit_score > prev[1]):
-                    best_parent[uid] = (p, unit_score)
-    
-            # keep only units best assigned to this parent
-            parent_scores[p] = [(u, sc) for (u, sc) in unit_rows if best_parent.get(u, (None, -1e9))[0] == p]
-    
-        # ---- quotas across parents (equal or proportional) ----
-        n_avail = sum(len(v) for v in parent_scores.values())
-        n_enriched = min(n_enr, n_avail)
-        quotas: dict[str, int] = {}
-        if equalize:
-            base = (n_enriched // max(1, len(parents)))
-            quotas = {p: base for p in parents}
-            rem = n_enriched - base * len(parents)
-            if rem > 0:
-                order = sorted(parents, key=lambda L: len(parent_scores.get(L, [])), reverse=True)
-                for p in order[:rem]:
-                    quotas[p] = quotas.get(p, 0) + 1
-        else:
-            import math
-            tot = max(1, sum(len(parent_scores.get(p, [])) for p in parents))
-            raw = {p: (len(parent_scores.get(p, [])) * n_enriched) / tot for p in parents}
-            quotas = {p: int(math.floor(v)) for p, v in raw.items()}
-            give = n_enriched - sum(quotas.values())
-            if give > 0:
-                frac = sorted(((p, raw[p] - quotas[p]) for p in parents), key=lambda x: x[1], reverse=True)
-                for p, _ in frac[:give]:
-                    quotas[p] += 1
-    
-        # ---- pick by CE score within each parent; merge with uniform; top up uniformly if needed ----
-        chosen, used = [], set(uniform_units)
-        for p in parents:
-            pool = sorted(parent_scores.get(p, []), key=lambda t: t[1], reverse=True)
-            need = int(quotas.get(p, 0))
-            for (u, _sc) in pool:
-                if u in used: 
-                    continue
-                chosen.append(u); used.add(u)
-                if len(chosen) >= n_enriched or sum(1 for x in chosen if best_parent[x][0] == p) >= need:
-                    break
-    
-        merged = (uniform_units + chosen)
-        out, used = [], set()
-        for u in merged:
-            if u in used or u in ex:
-                continue
-            used.add(u); out.append(u)
-            if len(out) >= n_total: break
-        if len(out) < n_total:
-            rest = [u for u in all_units if u not in used]
-            if rest:
-                extra = rng.choice(rest, size=min(n_total - len(out), len(rest)), replace=False).tolist()
-                out.extend(extra)
-        return out[:n_total]
-
-
-    def sample_units_for_probe(self, exclude_units: Optional[set[str]] = None) -> List[str]:
-        import numpy as np
-        ex = set(exclude_units or [])
-        n_total = int(getattr(self.cfg, "n_probe_units", 100) or 100)
-        uids = [str(u) for u in sorted(self.repo.notes["unit_id"].unique().tolist()) if str(u) not in ex]
-        if not uids:
-            return []
-        rng = np.random.default_rng()
-        return rng.choice(uids, size=min(n_total, len(uids)), replace=False).tolist()
-
-
-    def probe_units_label_tree(self, enrich: bool, label_types: dict[str,str], 
-                               per_label_rules: dict[str,str], seed: Optional[int]=None,
-                               exclude_units: Optional[set[str]] = None) -> pd.DataFrame:
-        ex = set(exclude_units or [])
-        if enrich:
-            sample = self.sample_units_for_probe_enriched(per_label_rules, exclude_units=ex)
-        else: sample = self.sample_units_for_probe(exclude_units=ex)
-        rows = []
-        _progress_every = float(self.cfg.progress_min_interval_s or 1.0)
-        for uid in iter_with_bar(
-                step="LLM probe",
-                iterable=sample,
-                total=len(sample),
-                min_interval_s=_progress_every):
-            rows.extend(self.label_family_for_unit(uid, label_types, per_label_rules))
-
-        df = pd.DataFrame(rows)
-        return df
-
-# ------------------------------
-def per_label_kcenter(pool_df: pd.DataFrame, pooler: LabelAwarePooler, retriever: RAGRetriever, rules_map: Dict[str,str],
-                      per_label_quota: int, already_selected_vecs: Optional[np.ndarray]=None) -> pd.DataFrame:
-    out = []
-    for lid, grp in pool_df.groupby("label_id"):
-        if grp.empty: continue
-        k = min(per_label_quota, len(grp))
-        vecs = [
-            pooler.pooled_vector(
-                r.unit_id,
-                r.label_id,
-                retriever,
-                rules_map.get(r.label_id, ""),
-                context_builder=self.context_builder,
-            )
-            for r in grp.itertuples(index=False)
-        ]
-        V = np.vstack(vecs)
-        idxs = kcenter_select(V, k=k, seed_idx=None, preselected=already_selected_vecs)
-        sel = grp.iloc[idxs].copy()
-        sel = sel.reset_index(drop=True)
-        sel["kcenter_rank_per_label"] = list(range(1, len(sel)+1))
-        out.append(sel)
-    return pd.concat(out, ignore_index=True) if out else pool_df.head(0)
-
-
-def merge_with_global_kcenter(per_label_selected: pd.DataFrame, pool_df: pd.DataFrame,
-                              pooler: LabelAwarePooler, retriever: RAGRetriever, rules_map: Dict[str,str],
-                              target_n: int,
-                              already_selected_vecs: Optional[np.ndarray]=None) -> pd.DataFrame:
-    if per_label_selected.empty and pool_df.empty: return per_label_selected
-    selected = per_label_selected.copy()
-    cnt = Counter(selected["label_id"].tolist()) if not selected.empty else Counter()
-    remaining_needed = max(0, target_n - len(selected))
-    if remaining_needed == 0: return selected
-    if not selected.empty:
-        rem = pool_df.merge(selected[["unit_id","label_id"]], on=["unit_id","label_id"], how="left", indicator=True)
-        rem = rem[rem["_merge"]=="left_only"].drop(columns=["_merge"])
-    else:
-        rem = pool_df.copy()
-    if rem.empty: return selected
-
-    vecs = [
-        pooler.pooled_vector(
-            r.unit_id,
-            r.label_id,
-            retriever,
-            rules_map.get(r.label_id, ""),
-            context_builder=context_builder,
-        )
-        for r in rem.itertuples(index=False)
-    ]
-    V = np.vstack(vecs)
-
-    preV = None
-    if already_selected_vecs is not None and already_selected_vecs.size:
-        preV = already_selected_vecs
-    if not selected.empty:
-        Vsel = [
-            pooler.pooled_vector(
-                r.unit_id,
-                r.label_id,
-                retriever,
-                rules_map.get(r.label_id, ""),
-                context_builder=context_builder,
-            )
-            for r in selected.itertuples(index=False)
-        ]
-        Vsel = np.vstack(Vsel); preV = Vsel if preV is None else np.vstack([preV, Vsel])
-
-    idx_order = kcenter_select(V, k=min(remaining_needed, len(rem)), seed_idx=None, preselected=preV)
-
-    chosen_rows = []
-    for rank_idx, i in enumerate(idx_order, start=1):
-        row = rem.iloc[i]
-        row = row.copy()
-        row["kcenter_rank_global"] = rank_idx
-        chosen_rows.append(row); cnt[row.label_id] += 1
-        if len(chosen_rows) >= remaining_needed: break
-
-    if not chosen_rows: return selected
-    chosen_df = pd.DataFrame([r._asdict() if hasattr(r,"_asdict") else dict(r) for r in chosen_rows])
-    return pd.concat([selected, chosen_df], ignore_index=True)
-
-
-def direct_uncertainty_selection(
-    probe_df: pd.DataFrame,
-    n_unc: int,
-    *,
-    label_col: str = "label_id",
-    uncertainty_col: str = "U",
-    higher_is_more_uncertain: bool = True,   # e.g., U = entropy → higher = more uncertain
-    select_most_certain: bool = False,       # flip to pick the most certain items
-    add_selection_reason: bool = True,
-) -> pd.DataFrame:
-    """
-    Fill purely by round-robin across labels:
-      1) Sort within each label by the criterion (uncertainty or certainty).
-      2) Build a FIFO queue per label.
-      3) Pop one from each label in turn until n_unc is reached (or queues empty).
-
-    Notes:
-      - NaN uncertainty sinks to the end for the chosen ordering.
-    """
-    # Basic guards
-    if probe_df is None or not isinstance(probe_df, pd.DataFrame) or probe_df.empty or n_unc <= 0:
-        return (probe_df.head(0).copy()
-                if isinstance(probe_df, pd.DataFrame) else pd.DataFrame())
-
-    df = probe_df.copy()
-    if label_col not in df.columns or uncertainty_col not in df.columns:
-        raise ValueError(f"probe_df must contain '{label_col}' and '{uncertainty_col}'")
-
-    # Normalize uncertainty column and handle NaNs so they sink.
-    u = pd.to_numeric(df[uncertainty_col], errors="coerce")
-
-    # asc_u=True means "smaller score is better".
-    # We want:
-    #   - Most-uncertain:  desc if higher_is_more_uncertain else asc
-    #   - Most-certain:    asc  if higher_is_more_uncertain else desc
-    asc_u = (select_most_certain == higher_is_more_uncertain)
-
-    # Fill NaNs so they sort to the end given asc/desc choice
-    u = u.fillna(np.inf if asc_u else -np.inf)
-
-    # Tiny jitter for stable tie-breaking (optional)
-    rng = np.random.default_rng()
-    u = u + (rng.random(len(u)) * 1e-9)
-
-    df["_score"] = u
-
-    # Build per-label queues: each label sorted by the criterion
-    # (We sort by label then by score so groupby preserves label order;
-    #  optionally shuffle label order 
-    per_label_sorted = df.sort_values([label_col, "_score"], ascending=[True, asc_u])
-    label_groups = list(per_label_sorted.groupby(label_col, sort=False))
-
-    # Convert to queues (row-index lists)
-    queues = {lab: grp.index.to_list() for lab, grp in label_groups}
-    labels_order = [lab for lab, _ in label_groups if len(queues.get(lab, [])) > 0]
-
-    # Optional: shuffle label start order for fairness when a seed is provided
-    rng.shuffle(labels_order)
-
-    # Round-robin selection
-    chosen_idx = []
-    need = int(n_unc)
-
-    while need > 0 and labels_order:
-        progressed = False
-        new_order = []
-        for lab in labels_order:
-            q = queues.get(lab, [])
-            if q:
-                chosen_idx.append(q.pop(0))
-                need -= 1
-                progressed = True
-                if q:  # keep label in the rotation if it still has items
-                    new_order.append(lab)
-            if need <= 0:
-                break
-        labels_order = new_order if progressed else []
-
-    if not chosen_idx:
-        out = df.head(0).copy()
-        if add_selection_reason and "selection_reason" not in out.columns:
-            out["selection_reason"] = ("model_certain_direct" if select_most_certain
-                                       else "model_uncertain_direct")
-        return out
-
-    chosen = df.loc[chosen_idx].copy()
-
-    # Decorate
-    reason = "model_certain_direct" if select_most_certain else "model_uncertain_direct"
-    if add_selection_reason and "selection_reason" not in chosen.columns:
-        chosen["selection_reason"] = reason
-
-    prefix = "certain" if select_most_certain else "uncertain"
-    # Per-label rank by the criterion (for inspection)
-    chosen[f"{prefix}_rank_per_label"] = (
-        chosen.sort_values([label_col, "_score"], ascending=[True, asc_u])
-              .groupby(label_col, sort=False).cumcount() + 1
-    )
-    # Global rank by the criterion
-    chosen[f"{prefix}_rank_global"] = (
-        chosen.sort_values("_score", ascending=asc_u)
-              .reset_index(drop=True).index + 1
-    )
-
-    # Return ordered by the chosen criterion (you could return in selection order if preferred)
-    out = (chosen.sort_values("_score", ascending=asc_u)
-                 .drop(columns=["_score"])
-                 .reset_index(drop=True))
-    return out
-
-
-
 class ActiveLearningLLMFirst:
     def __init__(
         self,
